test: more and refactor
