+++ test/fixtures/server-test/webpack.simple.config.js
@@ -4,8 +4,8 @@ const path = require('path');
 
 module.exports = {
   mode: 'development',
-  context: path.resolve(__dirname),
-  entry: './foo.js',
+  context: __dirname,
+  entry: './simple.js',
   output: {
     path: path.resolve(__dirname, '../../outputs/'),
   }
