+++ test/fixtures/server-test/webpack.warning.config.js
@@ -5,7 +5,7 @@ const path = require('path');
 module.exports = {
   mode: 'development',
   context: path.resolve(__dirname),
-  entry: './foo.js',
+  entry: './warning.js',
   output: {
     path: path.resolve(__dirname, '../../outputs/'),
   },
